requirements.txt
beautifulsoup4==4.12.3
email_validator==2.2.0
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
Flask_WTF==1.0.1
itsdangerous==2.1.2
lxml==4.9.3
python-dotenv==1.0.1
Requests==2.32.3
schedule==1.2.2
WTForms==3.0.1
